PFE_MVP/requirements.txt
yfinance==0.2.65
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
ta==0.11.0
scikit-learn==1.5.1
torch>=2.2,<2.3
safetensors==0.4.5
pydantic==2.8.2
pyyaml==6.0.2
typer==0.12.5
rich==13.8.1
python-dotenv==1.0.1
pytest==7.4.0
